+++ src/test/activation/excludeFiles.ls.test.ts
